requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.1
defusedxml==0.7.1
Django==3.1.1
django-admin-rangefilter==0.6.2
django-admin-sortable2==0.7.7
django-ckeditor==5.9.0
django-cors-headers==3.5.0
django-embed-video==1.3.3
django-extensions==3.1.3
django-filter==2.3.0
django-js-asset==1.2.2
django-nested-inline==0.4.2
django-phonenumber-field==5.0.0
django-rest-multiple-models==2.1.3
djangorestframework==3.11.1
djangorestframework-filters==1.0.0.dev0
djangorestframework-simplejwt==4.4.0
drf-extra-fields==3.0.3
drf-writable-nested==0.6.3
drf-yasg==1.17.1
drf-yasg2==1.19.4
easy-thumbnails==2.7.1
idna==2.10
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.4
phonenumbers==8.12.12
Pillow==7.2.0
progressbar2==3.53.1
psycopg2==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pyotp==2.4.1
pyparsing==2.4.7
python-magic==0.4.24
python-utils==2.4.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
